requirements.txt
click==6.7
elasticsearch==5.5.2
Flask==0.12.2
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
python-snappy==0.5.2
urllib3==1.22
Werkzeug==0.14.1
